pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-engineering-intelligence"
version = "1.0.0"
description = "Daily news research + deep synthesis pipeline for engineering intelligence"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["news", "data-engineering", "mkdocs", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "readability-lxml>=0.8.1",
    "newspaper3k>=0.2.8",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "PyYAML>=6.0.1",
    "markdownify>=0.11.6",
    "Jinja2>=3.1.2",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
dei = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
